+++ CHANGELOG.md
@@ -14,6 +14,9 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Vscode's events are not being properly unregistered
 - Fix translation selection be reverted when editing yml directly
 
+### Changed
+- Optimize zoom
+
 ## [0.0.9] - 2023-08-28
 
 ### Added