+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the "BetonQuest" extension will be documented in this fil
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [0.0.10] - TBD
+## [0.0.10] - 2023-08-31
 
 ### Added
 - Better supports for translation detection and selection